+++ modules/client-common/test/context.test.ts
@@ -36,7 +36,9 @@ describe("Context instances", () => {
     expect(context).toBeInstanceOf(TestContext);
     expect(context.network.name).toBe("homestead");
     expect(context.network.chainId).toBe(1);
-    expect(context.daoFactoryAddress).toBe(LIVE_CONTRACTS.homestead.daoFactory);
+    expect(context.daoFactoryAddress).toBe(
+      LIVE_CONTRACTS.homestead.daoFactoryAddress,
+    );
     expect(context.ensRegistryAddress).toBe(context.network.ensAddress);
     expect(context.gasFeeEstimationFactor).toBe(0.625);
     expect(context.web3Providers.length).toBe(0);
@@ -110,7 +112,9 @@ describe("Context instances", () => {
     expect(context).toBeInstanceOf(TestContext);
     expect(context.network.name).toBe("goerli");
     expect(context.network.chainId).toBe(5);
-    expect(context.daoFactoryAddress).toBe(LIVE_CONTRACTS.goerli.daoFactory);
+    expect(context.daoFactoryAddress).toBe(
+      LIVE_CONTRACTS.goerli.daoFactoryAddress,
+    );
     expect(context.ensRegistryAddress).toBe(context.network.ensAddress);
     expect(context.gasFeeEstimationFactor).toBe(0.625);
     expect(context.web3Providers.length).toBe(1);
@@ -131,8 +135,12 @@ describe("Context instances", () => {
     });
     expect(context.network.name).toBe("matic");
     expect(context.network.chainId).toBe(137);
-    expect(context.daoFactoryAddress).toBe(LIVE_CONTRACTS.matic.daoFactory);
-    expect(context.ensRegistryAddress).toBe(LIVE_CONTRACTS.matic.ensRegistry);
+    expect(context.daoFactoryAddress).toBe(
+      LIVE_CONTRACTS.matic.daoFactoryAddress,
+    );
+    expect(context.ensRegistryAddress).toBe(
+      LIVE_CONTRACTS.matic.ensRegistryAddress,
+    );
     expect(context.gasFeeEstimationFactor).toBe(0.625);
     expect(context.web3Providers.length).toBe(1);
     expect(context.ipfs.length).toBe(IPFS_NODES.matic.length);
@@ -156,8 +164,12 @@ describe("Context instances", () => {
     });
     expect(context.network.name).toBe("matic");
     expect(context.network.chainId).toBe(137);
-    expect(context.daoFactoryAddress).toBe(LIVE_CONTRACTS.matic.daoFactory);
-    expect(context.ensRegistryAddress).toBe(LIVE_CONTRACTS.matic.ensRegistry);
+    expect(context.daoFactoryAddress).toBe(
+      LIVE_CONTRACTS.matic.daoFactoryAddress,
+    );
+    expect(context.ensRegistryAddress).toBe(
+      LIVE_CONTRACTS.matic.ensRegistryAddress,
+    );
     expect(context.gasFeeEstimationFactor).toBe(0.625);
     expect(context.web3Providers.length).toBe(1);
     expect(context.ipfs.length).toBe(IPFS_NODES.matic.length);
@@ -195,7 +207,9 @@ describe("Context instances", () => {
     expect(context).toBeInstanceOf(TestContext);
     expect(context.network.name).toBe("matic");
     expect(context.network.chainId).toBe(137);
-    expect(context.daoFactoryAddress).toBe(LIVE_CONTRACTS.matic.daoFactory);
+    expect(context.daoFactoryAddress).toBe(
+      LIVE_CONTRACTS.matic.daoFactoryAddress,
+    );
     expect(context.ensRegistryAddress).toBe(ADDRESS_ONE);
     expect(context.gasFeeEstimationFactor).toBe(0.625);
     expect(context.web3Providers.length).toBe(0);
@@ -236,14 +250,14 @@ describe("Context instances", () => {
       provider.getNetwork().then((nw) => {
         expect(nw.chainId).toEqual(137);
         expect(nw.name).toEqual("matic");
-        expect(nw.ensAddress).toEqual(LIVE_CONTRACTS.matic.ensRegistry);
+        expect(nw.ensAddress).toEqual(LIVE_CONTRACTS.matic.ensRegistryAddress);
       })
     );
     expect(context.daoFactoryAddress).toEqual(
-      LIVE_CONTRACTS.matic.daoFactory,
+      LIVE_CONTRACTS.matic.daoFactoryAddress,
     );
     expect(context.ensRegistryAddress).toEqual(
-      LIVE_CONTRACTS.matic.ensRegistry,
+      LIVE_CONTRACTS.matic.ensRegistryAddress,
     );
   });
   it("Should create a context with baseGoerli as network and have the correct values", () => {
@@ -256,10 +270,10 @@ describe("Context instances", () => {
     expect(context.network.name).toBe("baseGoerli");
     expect(context.network.chainId).toBe(84531);
     expect(context.daoFactoryAddress).toBe(
-      LIVE_CONTRACTS.baseGoerli.daoFactory,
+      LIVE_CONTRACTS.baseGoerli.daoFactoryAddress,
     );
     expect(context.ensRegistryAddress).toBe(
-      LIVE_CONTRACTS.baseGoerli.ensRegistry,
+      LIVE_CONTRACTS.baseGoerli.ensRegistryAddress,
     );
     expect(context.gasFeeEstimationFactor).toBe(0.625);
     expect(context.web3Providers.length).toBe(1);
@@ -269,7 +283,9 @@ describe("Context instances", () => {
       provider.getNetwork().then((nw) => {
         expect(nw.chainId).toEqual(84531);
         expect(nw.name).toEqual("baseGoerli");
-        expect(nw.ensAddress).toEqual(LIVE_CONTRACTS.baseGoerli.ensRegistry);
+        expect(nw.ensAddress).toEqual(
+          LIVE_CONTRACTS.baseGoerli.ensRegistryAddress,
+        );
       });
     }
   });